refactor: change routes
